Refactor: modifyCalenderToCalendar
