+++ src/main/java/promisor/promisor/domain/calendar/dao/PersonalScheduleRepository.java
@@ -1,0 +1,23 @@
+package promisor.promisor.domain.calendar.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.transaction.annotation.Transactional;
+import promisor.promisor.domain.calendar.domain.PersonalCalendar;
+import promisor.promisor.domain.calendar.domain.PersonalSchedule;
+
+import java.util.List;
+
+@Transactional(readOnly = true)
+public interface PersonalScheduleRepository extends JpaRepository<PersonalSchedule, Long>{
+
+    @Query(value = "select * from personal_shedule ps join personal_calendar pc on ps.personal_calendar_id = pc.id\n" +
+            "join member m on pc.member_id = m.id\n" +
+            "where pc.member_id = :memberId",
+    nativeQuery = true)
+    List<PersonalSchedule> findAllByMember(@Param("memberId") Long id);
+
+    @Query("select ps from PersonalSchedule ps join ps.personalCalendar pc where pc = :pc")
+    List<PersonalSchedule> findAllByPersonalCalendar(@Param("pc") PersonalCalendar pc);
+}
