+++ src/main/java/promisor/promisor/domain/calendar/domain/DateStatusType.java
@@ -1,0 +1,5 @@
+package promisor.promisor.domain.calendar.domain;
+
+public enum DateStatusType {
+    IMPOSSIBLE, UNCERTAIN, POSSIBLE
+}
