+++ src/main/java/promisor/promisor/domain/calendar/dto/GetPersonalScheduleResponse.java
@@ -1,0 +1,21 @@
+package promisor.promisor.domain.calendar.dto;
+
+import lombok.Data;
+import promisor.promisor.domain.calendar.domain.DateStatusType;
+
+import java.util.List;
+
+@Data
+public class GetPersonalScheduleResponse {
+    private DateStatusType dateStatus;
+    private List<String> reason;
+
+    public GetPersonalScheduleResponse(DateStatusType dateStatus, List<String> reason) {
+        this.dateStatus = dateStatus;
+        this.reason = reason;
+    }
+
+    public GetPersonalScheduleResponse(String dateStatus) {
+        this.dateStatus = DateStatusType.valueOf(dateStatus);
+    }
+}
