+++ src/main/java/promisor/promisor/domain/calendar/dto/PersonalCalendarStatusEditRequest.java
@@ -1,0 +1,9 @@
+package promisor.promisor.domain.calendar.dto;
+
+import lombok.Data;
+
+@Data
+public class PersonalCalendarStatusEditRequest {
+    String date;
+    String status;
+}
