+++ src/main/java/promisor/promisor/domain/calendar/dto/RegisterDateDto.java
@@ -1,0 +1,12 @@
+package promisor.promisor.domain.calendar.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Getter
+@NoArgsConstructor
+public class RegisterDateDto {
+    private String date;
+    private String status;
+}
