+++ src/main/java/promisor/promisor/domain/calendar/exception/DateEmptyException.java
@@ -1,0 +1,8 @@
+package promisor.promisor.domain.calendar.exception;
+
+import promisor.promisor.global.error.ErrorCode;
+import promisor.promisor.global.error.exception.EntityNotFoundException;
+
+public class DateEmptyException extends EntityNotFoundException {
+    public DateEmptyException() {super(ErrorCode.DATE_EMPTY);}
+}
