+++ src/main/java/promisor/promisor/domain/calendar/exception/StatusEmptyException.java
@@ -1,0 +1,10 @@
+package promisor.promisor.domain.calendar.exception;
+
+import promisor.promisor.global.error.ErrorCode;
+import promisor.promisor.global.error.exception.EntityNotFoundException;
+
+public class StatusEmptyException extends EntityNotFoundException {
+    public StatusEmptyException(){
+        super(ErrorCode.STATUS_EMPTY);
+    }
+}
