--- src/main/java/promisor/promisor/domain/calender/domain/DateStatusType.java
@@ -1,5 +1,0 @@
-package promisor.promisor.domain.calender.domain;
-
-public enum DateStatusType {
-    IMPOSSIBLE, UNCERTAIN, POSSIBLE
-}
