--- src/main/java/promisor/promisor/domain/calender/dto/GetTeamCalendarStatusResponse.java
@@ -1,17 +1,0 @@
-package promisor.promisor.domain.calender.dto;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class GetTeamCalendarStatusResponse {
-
-    private final LocalDate date;
-    private final String dateStatus;
-
-    public GetTeamCalendarStatusResponse(LocalDate date, String dateStatus) {
-        this.date = date;
-        this.dateStatus = dateStatus;
-    }
-}
