--- src/main/java/promisor/promisor/domain/calender/dto/ModifyStatusDto.java
@@ -1,15 +1,0 @@
-package promisor.promisor.domain.calender.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@NoArgsConstructor
-public class ModifyStatusDto {
-    @NotBlank(message = "날짜를 입력하세요.")
-    String date;
-    @NotBlank(message = "날짜상태를 입력하세요.")
-    String status;
-}
