--- src/main/java/promisor/promisor/domain/calender/dto/PersonalBanDateStatusEditRequest.java
@@ -1,9 +1,0 @@
-package promisor.promisor.domain.calender.dto;
-
-import lombok.Data;
-
-@Data
-public class PersonalBanDateStatusEditRequest {
-    String date;
-    String status;
-}
