--- src/main/java/promisor/promisor/domain/calender/dto/RegisterDateDto.java
@@ -1,12 +1,0 @@
-package promisor.promisor.domain.calender.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Getter
-@NoArgsConstructor
-public class RegisterDateDto {
-    private String date;
-    private String status;
-}
