--- src/main/java/promisor/promisor/domain/calender/exception/DateEmptyException.java
@@ -1,8 +1,0 @@
-package promisor.promisor.domain.calender.exception;
-
-import promisor.promisor.global.error.ErrorCode;
-import promisor.promisor.global.error.exception.EntityNotFoundException;
-
-public class DateEmptyException extends EntityNotFoundException {
-    public DateEmptyException() {super(ErrorCode.DATE_EMPTY);}
-}
