--- src/main/java/promisor/promisor/domain/calender/exception/PersonalBanDateNotFoundException.java
@@ -1,8 +1,0 @@
-package promisor.promisor.domain.calender.exception;
-
-import promisor.promisor.global.error.ErrorCode;
-import promisor.promisor.global.error.exception.EntityNotFoundException;
-
-public class PersonalBanDateNotFoundException extends EntityNotFoundException{
-    public PersonalBanDateNotFoundException() {super(ErrorCode.BANDATE_NOT_FOUND);}
-}
