--- src/main/java/promisor/promisor/domain/calender/exception/ReasonEmptyException.java
@@ -1,9 +1,0 @@
-package promisor.promisor.domain.calender.exception;
-
-import promisor.promisor.global.error.ErrorCode;
-import promisor.promisor.global.error.exception.EntityNotFoundException;
-
-public class ReasonEmptyException extends EntityNotFoundException {
-    public ReasonEmptyException() {super(ErrorCode.REASON_EMPTY);}
-}
-
