--- src/main/java/promisor/promisor/domain/calender/exception/RegisteredException.java
@@ -1,8 +1,0 @@
-package promisor.promisor.domain.calender.exception;
-
-import promisor.promisor.global.error.ErrorCode;
-import promisor.promisor.global.error.exception.EntityNotFoundException;
-
-public class RegisteredException extends EntityNotFoundException {
-    public RegisteredException(){super(ErrorCode.INVALID_INPUT_VALUE);}
-}
