--- src/main/java/promisor/promisor/domain/calender/exception/StatusEmptyException.java
@@ -1,10 +1,0 @@
-package promisor.promisor.domain.calender.exception;
-
-import promisor.promisor.global.error.ErrorCode;
-import promisor.promisor.global.error.exception.EntityNotFoundException;
-
-public class StatusEmptyException extends EntityNotFoundException {
-    public StatusEmptyException(){
-        super(ErrorCode.STATUS_EMPTY);
-    }
-}
