--- src/main/java/promisor/promisor/domain/calender/exception/WrongAccess.java
@@ -1,8 +1,0 @@
-package promisor.promisor.domain.calender.exception;
-
-import promisor.promisor.global.error.ErrorCode;
-import promisor.promisor.global.error.exception.EntityNotFoundException;
-
-public class WrongAccess extends EntityNotFoundException {
-    public WrongAccess(){super(ErrorCode.ACCESS_DENIED);}
-}
